Bugfix: Incomplete reception of response messages
For GNTP response messages, only the first package would be received and
parsed. In practice, this would usually cut off the (mandatory)
'Response-Action' header, which would arrive in the second package.
This, in turn, would cause GNTPOK.validate() to throw an exception
because of the missing header.

The correct behaviour is to continue reading from the socket until the
response ends with <CRLF><CRLF>, indicating the proper end of the
message. As the standard states:
> A GNTP request must end with <CRLF><CRLF> (two blank lines). This
> indicates the end of the request to the receiver.