Initial attempt at hooks for subscriptions
In order to better backwards support the old library, instead of adding
new parameters to the register and notify methods, add *_hook methods
that can be overridden to add additional headers.

Move the add_origin_info() method out of the GNTP base class to the
GrowlNotifier class where it can be more easily overridden if needed.