Remove code duplication from gntp.config
By taking advantage of **kwargs we can remove a lot of the code
duplication. To support this, we add a notifierFactory method to our
mini method though we have to put the mini method at the bottom of the
source file now, to ensure GrowlNotifier is defined.

Fixes #33. Thanks @xdgc