Fix buffer class to ignore binary data better
Fix our buffer class to be better aware of what is being written.
If we're writing a header, make sure we're writing two byte strings
properly
If we're writing just a single line, make sure it's a byte string that
ends in our file marker.
Inherit from Python3's BytesIO so that we leave binary data properly
alone